backend/requirements.txt
fastapi==0.110.1
uvicorn==0.22.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.11
asyncpg==0.28.0
pydantic==2.0.3
redis==5.0.0
celery==5.3.4
openai==0.27.7
langchain==0.1.7